requirements.txt
numpy
opencv-python
packaging
pillow
pytesseract
